Implemented wish #41467: Added possibility to hide files with status "Unknown" by extending the current option "Hide Up-To-Date Files". This option is now called "Hide Unmodified Files".
The reason for this is that the "Hide Up-To-Date Files" and a new
"Hide Unknown Files" would have served a very similar purpose
(to only show files that are modified in the file view) just for a different
repository state.

Lets see if people like it, even though "unmodified file" isn't 100 % correct
description for the status "Unknown".

CCMAIL: 41467-done@bugs.kde.org

svn path=/trunk/kdesdk/cervisia/; revision=312135