Fix BR#82853:
This fixes the version selection buttons in the resolve dialog (regression introduced with KDE 3.2.3). Unfortunately the patch also re-introduces the Bug 74903 (see #82853 for the reasons why this is acceptable).

Please help me to test this thoroughly, so we can backport it to the branch and inform the packager about the fix. Ideas about how to also fix #74903 are very welcome.

CCMAIL: cervisia@kde.org

svn path=/trunk/kdesdk/cervisia/; revision=321284