Updating the import and checkout sections, adding undocumented options.
Added some basic information, binary x text files, branches, etc... This may be a little redundant, but this is the beginning of the help file and we don't want to scare readers away, so better get a little ahead :)

Changed from CVS command focus to Cervisia dialog focus. But I kept the CVS command issued, it is a useful reference.

svn path=/trunk/kdesdk/doc/cervisia/; revision=327369